+++ src/index.ts
@@ -7,6 +7,14 @@ import req from "./request"
 import {getDiscovery, getDiscoveryWithGatewayUrl} from "./discovery"
 import {createGetOpenIdConfig} from "./oidc-config"
 import type {ApiClientConfig} from "./schema/config"
+import type {PkceAuthoriseOptions, PkceExchangeOptions} from "./pkce"
+import type {GetAuthorizeUrlOptions, GetAuthorizeUrlUsingPKCEOptions} from "./get-auth-urls"
+import type {
+  ExchangeCodeForTokensOptions,
+  ExchangeCodeForTokensUsingPKCEOptions,
+  LocalParams,
+  ParamsFromCallback,
+} from "./exchange-code-for-token"
 const DEFAULT_TIMEOUT = 60000
 const DEFAULT_OIDC_CACHE_TTL_MS = 3600000 // 1 hour
 
@@ -157,4 +165,12 @@ export {
   Users,
   ApiClientConfig,
   Moneyhub,
+  PkceAuthoriseOptions,
+  PkceExchangeOptions,
+  GetAuthorizeUrlOptions,
+  GetAuthorizeUrlUsingPKCEOptions,
+  ExchangeCodeForTokensOptions,
+  ExchangeCodeForTokensUsingPKCEOptions,
+  LocalParams,
+  ParamsFromCallback,
 }